Refactor Federation classes. Remove all unnecessary overridden methods
