+++ rskj-core/src/main/java/co/rsk/peg/ErpFederation.java
@@ -1,9 +1,7 @@
 package co.rsk.peg;
 
-import co.rsk.bitcoinj.core.Address;
 import co.rsk.bitcoinj.core.BtcECKey;
 import co.rsk.bitcoinj.core.NetworkParameters;
-import co.rsk.bitcoinj.script.ErpFederationRedeemScriptParser;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.bitcoinj.script.ScriptBuilder;
 import co.rsk.peg.utils.EcKeyUtils;
@@ -17,16 +15,18 @@ import org.ethereum.config.blockchain.upgrades.ConsensusRule;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ErpFederation extends Federation {
-    private static final byte[] ERP_TESTNET_REDEEM_SCRIPT_BYTES = Hex.decode("6453210208f40073a9e43b3e9103acec79767a6de9b0409749884e989960fee578012fce210225e892391625854128c5c4ea4340de0c2a70570f33db53426fc9c746597a03f42102afc230c2d355b1a577682b07bc2646041b5d0177af0f98395a46018da699b6da210344a3c38cd59afcba3edcebe143e025574594b001700dec41e59409bdbd0f2a0921039a060badbeb24bee49eb2063f616c0f0f0765d4ca646b20a88ce828f259fcdb955670300cd50b27552210216c23b2ea8e4f11c3f9e22711addb1d16a93964796913830856b568cc3ea21d3210275562901dd8faae20de0a4166362a4f82188db77dbed4ca887422ea1ec185f1421034db69f2112f4fb1bb6141bf6e2bd6631f0484d0bd95b16767902c9fe219d4a6f5368ae");
+public abstract class ErpFederation extends Federation {
     private static final Logger logger = LoggerFactory.getLogger(ErpFederation.class);
+
+    protected static final byte[] ERP_TESTNET_REDEEM_SCRIPT_BYTES = Hex.decode("6453210208f40073a9e43b3e9103acec79767a6de9b0409749884e989960fee578012fce210225e892391625854128c5c4ea4340de0c2a70570f33db53426fc9c746597a03f42102afc230c2d355b1a577682b07bc2646041b5d0177af0f98395a46018da699b6da210344a3c38cd59afcba3edcebe143e025574594b001700dec41e59409bdbd0f2a0921039a060badbeb24bee49eb2063f616c0f0f0765d4ca646b20a88ce828f259fcdb955670300cd50b27552210216c23b2ea8e4f11c3f9e22711addb1d16a93964796913830856b568cc3ea21d3210275562901dd8faae20de0a4166362a4f82188db77dbed4ca887422ea1ec185f1421034db69f2112f4fb1bb6141bf6e2bd6631f0484d0bd95b16767902c9fe219d4a6f5368ae");
 
     protected final List<BtcECKey> erpPubKeys;
     protected final long activationDelay;
     protected final ActivationConfig.ForBlock activations;
-
     protected Script standardRedeemScript;
+    // add erpRedeemScript? is it necessary?
     protected Script standardP2SHScript;
+    // add erpP2SHScript?
 
     public ErpFederation(
         List<FederationMember> members,
@@ -57,90 +57,14 @@ public class ErpFederation extends Federation {
         return activationDelay;
     }
 
-    @Override
-    public Script getRedeemScript() {
-        if (!activations.isActive(ConsensusRule.RSKIP284) &&
-            btcParams.getId().equals(NetworkParameters.ID_TESTNET)) {
-            logger.debug("[getRedeemScript] Returning hardcoded redeem script");
-            return new Script(ERP_TESTNET_REDEEM_SCRIPT_BYTES);
-        }
+    public abstract Script getStandardRedeemScript();
 
-        if (redeemScript == null) {
-            logger.debug("[getRedeemScript] Creating the redeem script from the keys");
-            redeemScript = activations.isActive(ConsensusRule.RSKIP293) ?
-                ErpFederationRedeemScriptParser.createErpRedeemScript(
-                    ScriptBuilder.createRedeemScript(getNumberOfSignaturesRequired(), getBtcPublicKeys()),
-                    ScriptBuilder.createRedeemScript(erpPubKeys.size() / 2 + 1, erpPubKeys),
-                    activationDelay
-                ) :
-                ErpFederationRedeemScriptParser.createErpRedeemScriptDeprecated(
-                    ScriptBuilder.createRedeemScript(getNumberOfSignaturesRequired(), getBtcPublicKeys()),
-                    ScriptBuilder.createRedeemScript(erpPubKeys.size() / 2 + 1, erpPubKeys),
-                    activationDelay
-                );
-        }
-
-        return redeemScript;
-    }
-
-    @Override
-    public Script getStandardRedeemScript() {
-        if (standardRedeemScript == null) {
-            standardRedeemScript = ErpFederationRedeemScriptParser.extractStandardRedeemScript(
-                getRedeemScript().getChunks()
-            );
-        }
-        return standardRedeemScript;
-    }
-
-    @Override
-    public Script getP2SHScript() {
-        if (p2shScript == null) {
-            p2shScript = ScriptBuilder.createP2SHOutputScript(getRedeemScript());
-        }
-
-        return p2shScript;
-    }
-
-    @Override
     public Script getStandardP2SHScript() {
         if (standardP2SHScript == null) {
             standardP2SHScript = ScriptBuilder.createP2SHOutputScript(getStandardRedeemScript());
         }
 
         return standardP2SHScript;
-    }
-
-    @Override
-    public Address getAddress() {
-        if (address == null) {
-            address = Address.fromP2SHScript(btcParams, getP2SHScript());
-        }
-
-        return address;
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (this == other) {
-            return true;
-        }
-
-        if (other == null || this.getClass() != other.getClass()) {
-            return false;
-        }
-
-        ErpFederation otherErpFederation = (ErpFederation) other;
-
-        return this.getNumberOfSignaturesRequired() == otherErpFederation.getNumberOfSignaturesRequired() &&
-            this.getSize() == otherErpFederation.getSize() &&
-            this.getCreationTime().equals(otherErpFederation.getCreationTime()) &&
-            this.creationBlockNumber == otherErpFederation.creationBlockNumber &&
-            this.btcParams.equals(otherErpFederation.btcParams) &&
-            this.members.equals(otherErpFederation.members) &&
-            this.getRedeemScript().equals(otherErpFederation.getRedeemScript()) &&
-            this.erpPubKeys.equals(otherErpFederation.erpPubKeys) &&
-            this.activationDelay == otherErpFederation.activationDelay;
     }
 
     @Override
